Summary: Adding more teuchos tests/examples and fixing TRILINOS_ADD_TEST(...)
Here I added several more example and test executables.

I also fixed the use of PASS_REGULAR_EXPRESSION.  It was not tested before but
not it is.

The testing support is very clean and I like what I see so far.

I built and tested the serial and MPI version with cmake.