SUMMARY:   Merging back changes to release branch
     BUGS:      3874, 3940

     CHECKLIST:

     1)   Did you add any source files? No.

     2)   Did you update or change the build system?  No.

     3) Did you rerun the test suite on at least the packages that you
        changed?  Yes.

     COMMENTS:

     A new accessor has been added to the solver manager class to get the
number of iterations.  This is being merged back from the release branch.
Also, the use of the assert() function and passing it a non-const operation
has been removed from the Belos code.  If Belos is compiled with -DNDEBUG
then the assert functions are removed and hence the operations performed
therein.  If those operations are necessary, say, to FillComplete() a matrix
or Compute() a preconditioner, then those operations are not performed and
the examples/tests crash.

(checklist version 2007-06-20)