Summary: More progress on automated dependency handling
I have made more progress on automated interpackage dependency handling.

Here is what I have done:

(*) Added tracing for what enables and disables are changed

(*) Added a macro TRILINOS_ADD_OPTION(...) to add not just a user option but
also the name of a cmake varaible that will be turned on that is to be used to
define a macro in the XXX_config.h file.  Currenly, this macro is only getting
set for optional package dependent packages.  I need a more general system to
handle all uses of this.

(*) I removed the option Epetra_ENABLE_FORTRAN from Epetra and replaced it
with Trilinos_ENABLE_FORTRAN defined at the global level.  This is consistent
with how it is done in the current autotools system with the global
--enable-fortran option that any package in Trilinos can hook into.

(*) All of the optional dependent packages in
${PACKAGE_NAME}_LIB_OPTIONAL_DEP_PACKAGES result in the user option
${PACKAGE_NAME}_ENABLE_${DEP_PACAKGE_NAME} being added to the external cache.
By default, this is left as empty "".  Howeever, when
Trilinos_ENABLE_${PACKAGE_NAME}=ON and Trilinos_ENABLE_${DEP_PACAKGE_NAME}=ON,
this gets turned own.  The variable HAVE_${PACKAGE_NAME}_${DEP_PACAKGE_NAME}
(where the names ${PACKAGE_NAME} and ${DEP_PACAKGE_NAME} are upper cased) is
also set which results in these getting automatically set in the XXX_config.h
file.

(*) Cleaned up RBGen's CMake files.  However, since RBGen very strange in that
it creates a single exectuable and all tests are based on that executable,
there are still no tests yet for RBGen.  Someone will have to add the CMake
logic to build the RBGen executable and define the tests.  A lot of RBGen
files are modified as I tried to add real tests but I gave up when I saw what
the real deal was.  As far as I know, RBGen is the only Trilinos package that
defines an exectuble as its primary product instead of a library.

(*) I changed all of the ctest scripts in Trilinos/cmake/ctest to enable all
packages and adjust the other options.  I am not sure how to test these
exactly.  I will need to talk with Danny and/or Tim about this.

I have tested this on my linux machine both with serial debug and mpi
optimized with cmake.