Summary: Adding unit tests for dependency handling logic
I have added some unit tests for the dependency handling capabilities.  The
current set of packages that I have to work with is pretty small but I think
there is enough here to keep the capability safe.  These unit tests are
enabled by default on non-windows systems.  I did this because I am using a
unix shell script and the chmod command.  I don't think it is important to run
these unit tests on MS Windows.

A few other things to note:

(*) Generate_Script_File_Target.cmake: Defines a macro
GENERATE_SCRIPT_FILE_TARGET(...) that makes in really easy to define a script
target file that will get gnerated just given a script file name and a string
containing the script.

(*) Multiline_Set.cmake: Created a nice little helper script for creating a
single SET(...) call using an array of string arguments.  This makes it much
easier to format a multi-line string and still use indentation correctly."
The only problem is that you should avoid ';' in the string text or it will
have a fit because it thinks it is an array.  CMake is kind of messed up this
way.

I tested this with both serial debug and mpi optimized with cmake.