SUMMARY:   added configure-time test of BLAS complex dot products, solution
     BUGS:      4261

     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  yes
     2.a) Did you remember to do a bootstrap and do a 'cvs add' on the
          Makefile.am files and the Makefile.in files?  yes
     2.b) [Optional but recommended] Did you blow away the libraries and
          executables and then rebuild?  yes

     3) Did you rerun the test suite on at least the packages that you
        changed?  yes

     COMMENTS:

The way that Fortran-compiled BLAS routines return a complex number was causing problems when called from C/C++.
I added a configure-time test to determine
a) whether this was a problem
b) if so, whether it could be addressed.
These tests are currently disabled, requiring --enable-teuchos-dotcheck to be enabled. If so, the wrapper for the
third-party BLAS routines CDOTC and ZDOTC by Teuchos::BLAS::DOT() are modified.

Currently, only Tpetra stresses this "bug".

(checklist version 2007-06-20)