Summary: Fixing case of the option Teuchos_ENABLE_Fortran
Bugs: 4239

Here, I changed some instances of the name Trilinos_ENABLE_FORTRAN to
Trilinos_ENABLE_Fortran.  The option should be called
Trilinos_ENABLE_Fortran because CMake uses th case 'Fortran' for
options like CMAKE_Fortran_FLAGS and others.  This was causing my
dependency checking tests godal to fail because it thinks that Fortran
is really enabled and tried to check the default fortran compiler
which does not work.

Once I check this in, the dependency tests for Fortran should work
again.

NOTE: Variable names in CMake are case sensitive even if command names
are not.  This is evil!

Tested on godel.sandia.gov with:

- CMake serial debug rebuild (all 227 tests ran passed)

- CMake mpi release rebuild (all 208 tests ran passed)