Added a couple more tests using the cmake infrastructure.
Cleaned up compiler warnings that the debug flags exposed on linux/gnu-4.2.4.

This commit also includes quite a bit of in-progress refactoring, but tests
are passing. This is long-overdue refactoring to clean up a couple APIs and
will, when finished, eliminate some old fei classes and replace them with
much-improved classes and functions that are more robust and easier to
debug.

More tests will be added soon to increase cmake test coverage.