Summary: Adjusting optional TPL enables to be consistent with bug 4298
Bugs: 4298

Here I moved the logic to turn on TPL_ENABLE_<TPL>=ON when
<PACKAGE>_ENABLE_<TPL>=ON after all of the other enables so that TPL support
would only be turned on for that package and not other packages.  This is what
was requested in bug 4298.

Now, the following use cases, as described in bug 4298, should all be
supported:

TPL_ENABLE_TPL1:BOOL=ON
 - TPL1 is enabled for all packages

PACKAGE1_ENABLE_TPL1:BOOL=ON
 - TPL1 is enabled for package1 only

TPL_ENABLE_TPL1:BOOL=ON
PACKAGE1_ENABLE_TPL1:BOOL=OFF
 - TPL1 is enabled for all packages except package1

I need better unit tests to verify these three behaviors but I am pretty
confident based on the unit tests that I have and of the simplicity of the
CMake code that they are satsified.

I only ran the dependency unit tests which all passed.