Summary: Made CMake code unit test platform independnet
I had stupidly put in a unit test that hard-coded a path on my machine in a
comparion check.  In its place, I put in a series of regular expression checks
that ensures that all of the major parts of the CMake script are being written
correctly.  If the scirpt does not run after this then that will be caught by
other unit tests.