Summary: Adding strong warning output for things that are explicit 'ON' set to 'OFF'
Here I have changed the CMake configure code to print a strong warning message
when something that was explicitly set to 'ON' is set to 'OFF' because a
required dependency turned it off.  For example, if you configure with:

  $ ./do-configure -DTrilinos_ENABLE_Teuchos=OFF -DTrilinos_ENABLE_RTOp=ON

you will now see the following warning message in the configure output:

***
*** WARNING: Setting Trilinos_ENABLE_RTOp=OFF which was 'ON' because RTOp has a required library dependence on disabled package Teuchos!
***

The reason that I did this was to make it more clear why a package fails to
configure because one of its required dependencies is disabled.  This should
make configure failures like this in the CDash output more clear.  Hopefully
people will notice this easier.

In general, if the CMake system is going to disable something that a user
explicitly enabled, the configure output should make this very clear and make
some noise.

I tested this by configuring all of Trilinos as:

  $ /do-configure -DTrilinos_ENABLE_ALL_PACKAGES=ON

Everything configured and the otput looked correct.